Now code is restructured each module in a separate file Walid, thankfully, added the code of the non-ancilla design of both, the oracle and the diffuser.
new:
- I resturctured the code and did some modifications, like deleting the
number of iterations argument, as it can be computed easily.

- I tried to make a function that applies n-toffoli gates, but in Walid's
design, it had similar, but slightly different design, so it was
implemented twice. one in the oracle function, and one in the diffuser
function.

- I added some tests in the test_grover.py file.

- There is a jupyter notebook to test and visualize the results that Walid
sent to me, and I just edited few lines in it (to fit new argurments)

---------------------------
to do next:
- do the computational complexity in different number of qubits and
  visualise them (easy)

- Build a good readme file

- Add noise (Omar and Walid should do / are doing this, right now)